+++ src/base/components/inputs/LanguageSelect.tsx
@@ -49,7 +49,7 @@ const LanguageSelectDialog = ({
     const languagesSortedBySelectState = useMemo(
         () =>
             toUniqueLanguageCodes([
-                ...tmpSelectedLanguages
+                ...selectedLanguages
                     .filter((language) => languages.includes(language))
                     .toSorted(languageSortComparator),
                 ...languages.toSorted(languageSortComparator),